--- sources/en.mangadotnet/res/source.json
@@ -1,12 +1,0 @@
-{
-	"info": {
-		"id": "en.mangadotnet",
-		"name": "Mangadotnet",
-		"version": 1,
-		"url": "https://mangadot.net",
-		"contentRating": 1,
-		"languages": [
-			"en"
-		]
-	}
-}
