+++ sources/multi.mangadotnet/Cargo.toml
