+++ sources/multi.mangadotnet/res/filters.json
