+++ sources/multi.mangadotnet/res/settings.json
