+++ sources/multi.mangadotnet/res/source.json
@@ -1,0 +1,53 @@
+{
+	"info": {
+		"id": "multi.mangadotnet",
+		"name": "Mangadotnet",
+		"version": 1,
+		"url": "https://mangadot.net",
+		"contentRating": 1,
+		"languages": [
+			"en",
+			"ar",
+			"bn",
+			"bg",
+			"my",
+			"zh-Hans",
+			"zh-Hant",
+			"cs",
+			"da",
+			"nl",
+			"fil",
+			"fi",
+			"fr",
+			"ka",
+			"de",
+			"el",
+			"he",
+			"hi",
+			"hu",
+			"id",
+			"it",
+			"ja",
+			"ko",
+			"la",
+			"lt",
+			"ms",
+			"mn",
+			"no",
+			"fa",
+			"pl",
+			"pt",
+			"pt-BR",
+			"ro",
+			"ru",
+			"es",
+			"es-419",
+			"sv",
+			"th",
+			"tr",
+			"uk",
+			"vi"
+		],
+		"minAppVersion": "0.8.0"
+	}
+}
