+++ sources/multi.mangadotnet/src/helpers.rs
